model の配置が悪かったので修正
