+++ app/views/todo/list.scala.html
@@ -3,7 +3,7 @@
  * to do sample project
  *
  *@
-@import model.Todo
+@import lib.model.Todo
 @(vv: model.ViewValueCommon, todos: Seq[Todo#EmbeddedId])
 
 @common.Default(vv){
